% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.tree}
\alias{plot.tree}
\title{Plots a discussion tree using nice parameters}
\usage{
\method{plot}{tree}(gtree, labels = c("name", "id"))
}
\arguments{
\item{gtree}{igraph object}

\item{labels}{type of label to be shown over vertices, if any}
}
\description{
Plots a discussion tree using nice parameters
}

